backend/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
decorator==4.4.2
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.23.5
Pillow==9.3.0
proglog==0.1.10
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.13
youtube-dl==2021.12.17
